Desktop/finalpfe/backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
flask-sqlalchemy==3.1.1
mysql-connector-python==8.2.0
bcrypt==4.1.1
pyjwt==2.8.0
openpyxl==3.1.2

